+++ app/components/features/header/ViewHeaderSearch.vue
@@ -101,6 +101,24 @@
                     :class="{ 'search__results-airports-match': exactAirportsMatch }"
                 >
                     <div
+                        v-if="searchResults.history"
+                        class="__info-sections search__results search__results--history"
+                    >
+                        <ui-block-title
+                            :bubble="searchResults.history.length"
+                            remove-margin
+                        >
+                            History
+                        </ui-block-title>
+                        <ui-text-block
+                            v-for="item in searchResults.history"
+                            :key="item"
+                            :bottom-items="[item]"
+                            is-button
+                            @click="search = item"
+                        />
+                    </div>
+                    <div
                         v-if="searchResults.flights && canSearchBy('flights')"
                         class="__info-sections search__results search__results--flights"
                     >
@@ -296,14 +314,16 @@ watch([search, opened], async ([val]) => {
     val = val.trim().toUpperCase();
 
     if (!val) {
-        searchResults.value = {};
+        searchResults.value = {
+            history: history.value,
+        };
         return;
     }
 
     const results: Partial<SearchResults> = {};
 
     history.value.unshift(val);
-    history.value = history.value.slice(0, 10);
+    history.value = Array.from(new Set(history.value.slice(0, 10)));
 
     exactAirportsMatch.value = false;
 
@@ -478,11 +498,9 @@ onMounted(() => {
     }
 
     &_window {
-        scrollbar-gutter: stable;
-
         position: absolute;
         z-index: 10;
-        top: 100%;
+        top: calc(100% - 2px);
         left: 0;
 
         overflow: auto;
@@ -491,9 +509,9 @@ onMounted(() => {
         gap: 16px;
 
         width: 100%;
-        min-height: 320px;
         max-height: 50vh;
-        padding: 16px;
+        padding: 8px;
+        border: 1px solid $darkGray200;
         border-bottom-right-radius: 8px;
         border-bottom-left-radius: 8px;