+++ app/composables/render/update/atc.ts
@@ -13,6 +13,7 @@ import { simawareCache } from '~/composables/render/airports';
 export const callsignSplitRegex = /_+/gm;
 
 let uirsMap: Record<string, VatSpyData['uirs'][0]> | undefined;
+const uirsMapByCallsign: Record<string, string[]> = {};
 
 const firsMap: Record<string, VatSpyData['firs'][0]> = {};
 const firsMapByIcao: Record<string, string[]> = {};
@@ -26,7 +27,7 @@ export interface FirFindResult {
 const setVatspyBoundaries = new Set<string>();
 
 async function filterFirsForList(list: string[] | undefined, callsign: string) {
-    if (!list) return [];
+    if (!list?.length) return [];
 
     const dataStore = useDataStore();
 
@@ -96,6 +97,57 @@ function findFirsForCallsign(callsign: string, prefix?: string) {
     ], callsign);
 }
 
+function findUirForCallsign(callsign: string, prefix: string) {
+    const list = uirsMapByCallsign[prefix] ?? [];
+    if (!list?.length) return undefined;
+
+    const result: {
+        uir: VatSpyData['uirs'][0];
+        symbols: number;
+        exact: boolean;
+    }[] = [];
+
+    let callsignSplit = callsign.split(callsignSplitRegex);
+    callsignSplit = callsignSplit.slice(0, callsignSplit.length - 1);
+
+    const callsignMiddle = callsignSplit.join('_');
+
+    let maxStart = 0;
+    const foundExact = false;
+
+    for (const item of list) {
+        const uir = uirsMap?.[item];
+        const exact = false;
+
+        if (!uir || !callsign.startsWith(uir.icao)) {
+            continue;
+        }
+
+        const word = '';
+
+        for (let i = 0; i < uir.icao.length; i++) {
+            if (!callsign.startsWith(uir.icao) && !callsign.startsWith(uir.icao)) break;
+        }
+
+        let length = word.length;
+
+        if (uir.icao === callsignMiddle) length = 100;
+
+        if (length < maxStart) {
+            continue;
+        }
+        maxStart = length;
+
+        result.push({
+            uir,
+            symbols: length,
+            exact,
+        });
+    }
+
+    return result.find(x => (!foundExact || x.exact) && x.symbols === maxStart);
+}
+
 function addSector(context: DataUpdateContext, sector: FirFindResult, controller: VatsimShortenedController | null, uir?: DataSector['uir']) {
     const sectorKey = !sector.fir
         ? sector.feature.properties.id
@@ -153,7 +205,12 @@ export async function updateControllers(context: DataUpdateContext) {
             uirsMap = {};
 
             for (const uir of dataStore.vatspy.value?.data.uirs ?? []) {
-                uirsMap[uir.icao.split(callsignSplitRegex)[0]] = uir;
+                const icao = uir.icao.split(callsignSplitRegex)[0];
+
+                uirsMap[uir.icao] = uir;
+
+                uirsMapByCallsign[icao] ??= [];
+                uirsMapByCallsign[icao].push(uir.icao);
             }
 
             for (const fir of dataStore.vatspy.value?.data.firs ?? []) {
@@ -274,16 +331,16 @@ export async function updateControllers(context: DataUpdateContext) {
         if (!isATIS && (controller.facility === facilities.CTR || controller.facility === facilities.FSS)) {
             if (!uirsMap) continue;
 
-            const uir = uirsMap[prefix];
+            const uir = findUirForCallsign(callsign, prefix);
 
             if (uir) {
                 let foundFir = false;
 
-                for (const fir of uir.firs.split(',')) {
+                for (const fir of uir.uir.firs.split(',')) {
                     const firs = await findFirsForCallsign(fir);
                     foundFir ||= !!firs.length;
 
-                    firs.forEach(x => addSector(context, x, controller, uir));
+                    firs.forEach(x => addSector(context, x, controller, uir.uir));
                 }
 
                 if (foundFir) continue;
@@ -368,7 +425,7 @@ export async function updateControllers(context: DataUpdateContext) {
             if (isApp && !feature && airport?.isPseudo) continue;
 
             if (!airport && feature) {
-                const key = feature?.properties.id + feature?.properties.prefix.join(',');
+                const key = feature?.properties.id + feature?.properties.prefix.map(x => x.split(callsignSplitRegex)[0]).join(',');
                 context.airportsAdded.add(key);
                 context.airports[key] ??= {
                     icao: validPrefix ?? feature.properties.id,